pyproject.toml
[tool.poetry]
name = "db_manager"
version = "0.1.0"
description = ""
authors = ["MaddoSaru <riquelme.madmonkey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mysql-connector-python = "^8.1.0"
python-dotenv = "^1.0.0"
simple-term-menu = "^1.6.1"
polars = "^0.19.3"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
